[Refactor - Data Size Update] #56
Updated the network data structures to be compatible with the new HidNpadButton definitions.

Fixed an alignment issues with MSVC where the HidNpadButton enum for Windows would be 32bit instead of 64bit.

Fixed mouse config structure alignment issues.

PC program now prints out the details of the command payload received from the Switch.

Updated vscode cpp properties to target the updated libraries.

+++ common/srp/controller/ControllerOptions.cpp
@@ -94,7 +94,7 @@ namespace controller
             {
                 auto secondsStr = entry->second;
                 auto const seconds = std::atof(secondsStr.c_str());
-                auto const nanoseconds = (uint32_t)timeutil::secondToNano(seconds);
+                auto const nanoseconds = timeutil::secondToNano(seconds);
                 data.homeButtonTriggerTime = std::clamp(nanoseconds, minHomeButtonTriggerTime, maxHomeButtonTriggerTime);
             }
             else
